requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
dask==2025.7.0
dask-geopandas==0.5.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fsspec==2025.7.0
geopandas==1.1.1
gps_helper==1.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.5
numpy==2.3.2
osmnx==2.0.5
packaging==25.0
pandas==2.3.1
partd==1.4.2
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.33.2
sgp4==2.24
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
toolz==1.0.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
